requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0

# Optional - for model optimization
# onnxruntime>=1.15.0
# optuna>=3.3.0  # Phase 3
